feat: :sparkles: Added settings and added save to navbar
So in this change we have added a simple but good settings ui, that will become a fully working settings menu in the future, it also comes with a sliding animation that we like, but might be changed in the future. Together with that we also added some local storage to the navbar, so users can refresh and the indicator stays on the selected tab

Currently the navbar has a bug on project/app open. If the user logged out while on the settings tab i wtill stay there, meaning if they reopen the app , they will be greated with the home screen, but the indicator will be settings/last used